Should be fixed

+++ code/calculator.py
@@ -1,4 +1,4 @@
 """A simple calculator module that provides basic operations."""
-def add(a: float, b: float) -> int:
+def add(a: float, b: float) -> float:
     """Add two numbers and return the result."""
     return a + b
